Add bonuses (#2)

+++ Assets/Scripts/BlockScript.cs
@@ -4,6 +4,7 @@ using UnityEngine.UI;
 public class BlockScript : MonoBehaviour
 {
     public GameObject textObject;
+
     Text textComponent;
     public int hitsToDestroy;
     public int points;
@@ -18,12 +19,12 @@ public class BlockScript : MonoBehaviour
                 textComponent.text = hitsToDestroy.ToString();
         }
         playerScript = GameObject.FindGameObjectWithTag("Player")
-            .GetComponent<PlayerScript>();
+                .GetComponent<PlayerScript>();
     }
 
     bool allModBlocksIsX()
     {
-        var blocksArray = GameObject.FindGameObjectsWithTag("Block");
+        var blocksArray = GameObject.FindGameObjectsWithTag("ModBlock");
 
         if (blocksArray.Length == 0)
             return false;
@@ -41,7 +42,7 @@ public class BlockScript : MonoBehaviour
 
     void deleteAllModBlocks()
     {
-        var blocksArray = GameObject.FindGameObjectsWithTag("Block");
+        var blocksArray = GameObject.FindGameObjectsWithTag("ModBlock");
         if (blocksArray.Length == 0)
             return;
 
@@ -58,34 +59,74 @@ public class BlockScript : MonoBehaviour
 
     private void OnCollisionEnter2D(Collision2D collision)
     {
-        hitsToDestroy--;
-        if (hitsToDestroy == 0)
+        print(collision.gameObject.GetComponent<BallScript>());
+        if (!collision.gameObject.GetComponent<BallScript>()) return;
+        hitsToDestroy -= collision.gameObject.GetComponent<BallScript>().damage;
+        if (gameObject.tag.Contains("Mod"))
+        {
+            if (textComponent != null)
+            {
+                if (textComponent.text.Equals("O"))
+                {
+                    textComponent.text = "X";
+                    if (allModBlocksIsX())
+                    {
+                        deleteAllModBlocks();
+                    }
+                }
+                else if (textComponent.text.Equals("X"))
+                {
+                    textComponent.text = "O";
+                }
+            }
+        }
+        else if (hitsToDestroy <= 0)
         {
             if (gameObject.name.Contains("Green"))
             {
-                playerScript.CreateBaseBonusObject(transform.position);
-            }
+                int bonusType;
+                if (playerScript.gameData != null)
+                {
+                    bonusType = playerScript.gameData.getBonusType();
+                }
+                else
+                {
+                    bonusType = 5;
+                }
+                print(bonusType);
+
+                switch (bonusType)
+                {
+                    case 0:
+                        print("NO SPAWN");
+                        break;
+                    case 1:
+                        print("SPAWN BASE");
+                        playerScript.CreateBaseBonusObject(transform.position);
+                        break;
+                    case 2:
+                        print("SPAWN FIRE");
+                        playerScript.CreateFireBonusObject(transform.position);
+                        break;
+                    case 3:
+                        print("SPAWN NORM");
+                        playerScript.CreateNormBonusObject(transform.position);
+                        break;
+                    case 4:
+                        print("SPAWN STEEL");
+                        playerScript.CreateSteelBonusObject(transform.position);
+                        break;
+                    default:
+                        break;
+                }
+            } 
             Destroy(gameObject);
             playerScript.BlockDestroyed(points);
         }
-        else if (textComponent != null)
+        else
         {
-            if (textComponent.text.Equals("O"))
-            {
-                textComponent.text = "X";
-                if (allModBlocksIsX())
-                {
-                    deleteAllModBlocks();
-                }
-            }
-            else if (textComponent.text.Equals("X"))
-            {
-                textComponent.text = "O";
-            }
-            else
-            {
+            if (textComponent != null)
                 textComponent.text = hitsToDestroy.ToString();
-            }
         }
     }
 }
